Replace Prop Drilling by useContext Hook
